% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockSim-package.R
\docType{package}
\name{clockSim-package}
\alias{clockSim-package}
\title{clockSim: Streamlined Simulation of Circadian Gene Networks}
\description{
A user-friendly workflow for simulating circadian clock gene networks. Despite decades of advances in modeling circadian clock dynamics, the lack of accessible tools for reproducible simulation workflows hinders the integration of computational modeling with experimental studies. 'clockSim' addresses this gap by providing models and helper functions with step-by-step vignettes. This package opens up system-level exploration of the circadian clock to wet-lab experimentalists, and future development will include additional clock architectures and other gene circuit models. Currently implemented models are based on Leloup and Goldbeter (1998) \doi{10.1177/074873098128999934}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yeyuan98/clockSim}
  \item Report bugs at \url{https://github.com/yeyuan98/clockSim/issues}
}

}
\author{
\strong{Maintainer}: Ye Yuan \email{yeyu@umich.edu} (\href{https://orcid.org/0000-0001-9641-9102}{ORCID}) [copyright holder]

}
\keyword{internal}
