% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getOdinGen}
\alias{getOdinGen}
\title{Get all clock model Odin generator objects}
\usage{
getOdinGen()
}
\value{
Named list of Odin generator R6 objects.
}
\description{
Refer to package documentation on preconfigured models in this package.
This function is intended for advanced usage only; normally calling
other helper functions will suffice for simulation.
}
\examples{
names(getOdinGen()) # All available models
vignette("clock-models", "clockSim")
vignette("noisy-LG-model", "clockSim") # Noise-incorporated model using SDE simulation
}
