% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_phase}
\alias{plot_phase}
\title{2-D phase potrait plot}
\usage{
plot_phase(df_result, x, y, time = NULL)
}
\arguments{
\item{df_result}{Data frame containing results to plot}

\item{x}{Column to plot as X-axis}

\item{y}{Column to plot as Y-axis}

\item{time}{Column to color the trajectory}
}
\value{
ggplot object of phase portrait
}
\description{
Provides convenience function to plot simulation trajectory of result from
one run in 2-D phase space.
}
\examples{
vignette("clock-models", "clockSim")
}
