% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockSim-package.R
\docType{package}
\name{clockSim-package}
\alias{clockSim-package}
\title{clockSim: Simulation of the Circadian Clock Gene Network}
\description{
A preconfigured simulation workflow for the circadian clock gene network.
}
\author{
\strong{Maintainer}: Ye Yuan \email{yeyu@umich.edu} (\href{https://orcid.org/0000-0001-9641-9102}{ORCID})

}
\keyword{internal}
