% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockSim-package.R
\name{clockSim}
\alias{clockSim}
\title{clockSim: Simulation of the Circadian Clock Gene Network.}
\description{
This package provides preconfigured circadian clock gene network simulation
models based on the Odin simulation engine.
}
\details{
Circadian clock is a paradigm for studying negative-positive feedback
mechanisms in Biology. Molecular clock, a cell-autonomous
transcriptional-translational feedback, is one of the best known gene
network oscillators.

Despite widespread interest on molecular clock research, numerical
simulation of the clock remains elusive to many if not most wet-lab
biologists, which hindered generation of novel hypotheses and efficient
exploration of the clock parameter space.

This package aims to provide a low-friction workflow for molecular clock
simulation, thereby unleashing the power of in silico exploration to more.
\subsection{Models}{

The mypackage functions ...
}
}
